Break up JibPluginIntegrationTest
